subsection{Sprint 4} Our fourth sprint involved the following set of requirements: \begin{enumerate} \item Adding support for negative numebrs on the graph \item Allowing the graph to change size dynamically with the window size \item Add support for boolean types \item Add support for conditional statements \tem Create a degrees to radians function \end{enumerate}
\subsubsection{Requirement 1: }
This additon was relatively simple. We change to code4 to draw to lines that crossed in the center to act as the axis, we also modified our mapping function to account for  the change in position of (0,0).
\subsubsection{Requirement 2: }
During our research for this requirement, we found that JPanel has built in getWidth() and getHeight() functions. This meant we could replace our hard-coded constants for these values in our calculations which meant anything drawn in the window would automatically scale as the window size was changed.
We also found out abou the Points2D library which is compatible with Java 2D, so we changed our code to handle points with this library rather than manually.
\subsubsection{Requirement 3: }
This is another literal type and the comparissons for them are binary operators that exist within the ast.
\subsubsection{Requirement 4: }
\subsubsection{Requirement 5: }